refactor: move PageHead from layout to ui
