+++ src/components/layouts/index.ts
@@ -1,9 +1,7 @@
-import PageHead from "./PageHead";
-import type { PageHeadProps } from "./PageHead";
 import PolicyPageLayout from "./PolicyPageLayout";
 import type { PolicyPageLayoutProps } from "./PolicyPageLayout";
 import StickyScrollLayout from "./StickyScrollLayout";
 import type { StickyScrollLayoutProps } from "./StickyScrollLayout";
 
-export { PageHead, PolicyPageLayout, StickyScrollLayout };
-export type { PageHeadProps, PolicyPageLayoutProps, StickyScrollLayoutProps };
+export { PolicyPageLayout, StickyScrollLayout };
+export type { PolicyPageLayoutProps, StickyScrollLayoutProps };
