+++ src/pages/career/[slug].tsx
@@ -3,14 +3,13 @@ import { useRouter } from "next/router";
 import { FC, ReactElement, useEffect, useRef, useState } from "react";
 import dynamic from "next/dynamic";
 
-import { PageHead } from "@/components/layouts";
-
 import { CareerPositionDetailsSection } from "@/components/career";
 
 import {
   BackToPreviousPageLink,
   ContentContainer,
   PageBase,
+  PageHead,
 } from "@/components/ui";
 import {
   getClickUpTaskQuery,
