+++ src/pages/get-access.tsx
@@ -1,9 +1,8 @@
 import { FC, ReactElement } from "react";
 import Image from "next/future/image";
 
-import { PageHead } from "@/components/layouts";
 import { MailchimpSignupForm } from "../components/forms/MailchimpSignupForm";
-import { PageBase } from "@/components/ui";
+import { PageBase, PageHead } from "@/components/ui";
 
 interface GetLayOutProps {
   page: ReactElement;
